backend/requirements.txt
fastapi~=0.110
uvicorn[standard]~=0.29
pydantic~=2.7
python-dotenv~=1.0
httpx~=0.27
openai>=1.35.0